Add addition and subtraction operators to locations
